requirements.txt
beautifulsoup4>=4.14.2
python-dotenv>=1.0.0
icecream>=2.1.8
langchain>=1.0.3
langchain-google-genai
pydantic>=2.0.0
seleniumbase>=4.44.7
requests